Rename: 공통 hook 경로 변경
